# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"RoleResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"is_system": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"is_system"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRoleResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleResponseDto"
					}
				},
				"required": [
					"message",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRoleDto": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 50
					},
					"description": {
						"type": "string",
						"maxLength": 255
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetRolesResponseDto": {
				"properties": {
					"roles": {
						"items": {
							"$ref": "#/components/schemas/RoleResponseDto"
						},
						"type": "array"
					}
				},
				"required": [
					"roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PermissionLevel": {
				"enum": [
					"READ",
					"WRITE",
					"DELETE",
					"ALL"
				],
				"type": "string"
			},
			"RoleDetailResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"is_system": {
						"type": "boolean"
					},
					"permissions": {
						"items": {
							"properties": {
								"permission": {
									"$ref": "#/components/schemas/PermissionLevel"
								},
								"feature": {
									"type": "string"
								},
								"feature_id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"permission",
								"feature",
								"feature_id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"is_system",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoleResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/RoleResponseDto"
					}
				},
				"required": [
					"message",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoleDto": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 50
					},
					"description": {
						"type": "string",
						"maxLength": 255
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SetPermissionsResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"role_id": {
						"type": "number",
						"format": "double"
					},
					"permissions": {
						"items": {
							"properties": {
								"permission": {
									"$ref": "#/components/schemas/PermissionLevel"
								},
								"feature": {
									"type": "string"
								},
								"feature_id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"permission",
								"feature",
								"feature_id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"role_id",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RolePermissionDto": {
				"properties": {
					"feature_id": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"permission": {
						"$ref": "#/components/schemas/PermissionLevel"
					}
				},
				"required": [
					"feature_id",
					"permission"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SetPermissionsDto": {
				"properties": {
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/RolePermissionDto"
						},
						"type": "array"
					}
				},
				"required": [
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Emirate": {
				"enum": [
					"ABU_DHABI",
					"DUBAI",
					"SHARJAH",
					"AJMAN",
					"UMM_AL_QUWAIN",
					"RAS_AL_KHAIMAH",
					"FUJAIRAH"
				],
				"type": "string"
			},
...
			"SignupResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"business_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"user_id",
					"business_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BusinessRegistrationDto": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 255
					},
					"vat_number": {
						"type": "string",
						"minLength": 15,
						"maxLength": 15
					},
					"tl_number": {
						"type": "string",
						"maxLength": 50
					},
					"industry_id": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"emirate": {
						"$ref": "#/components/schemas/Emirate"
					}
				},
				"required": [
					"name",
					"vat_number",
					"tl_number",
					"industry_id",
					"emirate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignupDto": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 100
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"phone": {
						"type": "string",
						"pattern": "^\\+?[1-9]\\d{1,14}$"
					},
					"password": {
						"type": "string",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"
					},
					"business": {
						"$ref": "#/components/schemas/BusinessRegistrationDto"
					}
				},
				"required": [
					"name",
					"email",
					"phone",
					"password",
					"business"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyOtpResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"is_complete": {
						"type": "boolean"
					},
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"is_complete"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OtpPurpose.SIGNUP": {
				"enum": [
					"signup"
				],
				"type": "string"
			},
			"OtpPurpose.RESET_PASSWORD": {
				"enum": [
					"reset_password"
				],
				"type": "string"
			},
			"VerifyOtpDto": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"code": {
						"type": "string",
						"pattern": "^\\d{6}$"
					},
					"purpose": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/OtpPurpose.SIGNUP"
							},
							{
								"$ref": "#/components/schemas/OtpPurpose.RESET_PASSWORD"
							}
						]
					}
				},
				"required": [
					"identifier",
					"code",
					"purpose"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResendOtpDto": {
				"properties": {
					"identifier": {
						"type": "string"
					},
					"purpose": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/OtpPurpose.SIGNUP"
							},
							{
								"$ref": "#/components/schemas/OtpPurpose.RESET_PASSWORD"
							}
						]
					}
				},
				"required": [
					"identifier",
					"purpose"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponseDto": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					},
					"user": {
						"properties": {
							"business": {
								"properties": {
									"name": {
										"type": "string"
									},
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"name",
									"id"
								],
								"type": "object"
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"business",
							"email",
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"access_token",
					"refresh_token",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginDto": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"password": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshResponseDto": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"refresh_token": {
						"type": "string"
					}
				},
				"required": [
					"access_token",
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshDto": {
				"properties": {
					"refresh_token": {
						"type": "string",
						"minLength": 1
					}
				},
				"required": [
					"refresh_token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordDto": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordDto": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"code": {
						"type": "string",
						"pattern": "^\\d{6}$"
					},
					"new_password": {
						"type": "string",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"
					}
				},
				"required": [
					"email",
					"code",
					"new_password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "einvoicebackend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/roles": {
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateRoleResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRolesResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"OWNER",
							"ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/roles/{id}": {
			"get": {
				"operationId": "GetRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleDetailResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"OWNER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateRoleResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/roles/{id}/permissions": {
			"put": {
				"operationId": "SetPermissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SetPermissionsResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetPermissionsDto"
							}
						}
					}
				}
			}
		},
...
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignupResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupDto"
							}
						}
					}
				}
			}
		},
		"/auth/verify-otp": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyOtpResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyOtpDto"
							}
						}
					}
				}
			}
		},
		"/auth/resend-otp": {
			"post": {
				"operationId": "ResendOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResendOtpDto"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto"
							}
						}
					}
				}
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RefreshResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshDto"
							}
						}
					}
				}
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordDto"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}